requirements/base.txt
djangorestframework==3.9.0
isort==4.3.4
flake8==3.6.0
